fix klee_benchmark.py

+++ klee_benchmark.py
@@ -19,7 +19,7 @@ class TestRunner(object):
             self.build_directory = builddir
         else:
             self.build_directory = "build"
-        
+
         self.klee_path = os.path.join(self.base_directory, self.build_directory, "bin", "klee")
         self.include_path = os.path.join(self.base_directory, "include")
         self.lib_path = os.path.join(self.base_directory, self.build_directory, 'lib')
@@ -30,7 +30,7 @@ class TestRunner(object):
             print(f"Coverage output directory ({self.output_dir}) already exists. To continue, move or remove the directory.")
             sys.exit(1)
 
-        os.mkdir(self.output_dir)            
+        os.mkdir(self.output_dir)
 
         self.keep_output = keep_output
         self.tmp = tempfile.mkdtemp()
@@ -44,7 +44,7 @@ class TestRunner(object):
             self.klee_output_dir = os.path.join(self.tmp, "klee_output")
 
         os.mkdir(self.klee_output_dir)
-        
+
         self.timeout = timeout
 
 
@@ -52,16 +52,16 @@ class TestRunner(object):
         print(f"Compiling {source}")
         path_to_clang = os.path.join(self.llvm,"bin","clang") if self.llvm else "clang"
         path_to_llvm_link = os.path.join(self.llvm,"bin","llvm-link") if self.llvm else "llvm-link"
-        instance_dir = os.path.join(self.tmp,os.path.splitext(source)[0])
+        instance_dir = os.path.join(self.output_dir, os.path.splitext(source)[0])
         os.mkdir(instance_dir)
         shutil.copy(os.path.join(self.source_directory, source), instance_dir)
         source = os.path.join(instance_dir, source)
-        
+
         compiler_options = ["-O0", "-Xclang", "-disable-O0-optnone"]
         compiled_file = os.path.join(instance_dir, os.path.basename(source) + '.bc')
         cmd = [path_to_clang, "-I", self.include_path, "-c", "-Wno-everything",  "-g", "-emit-llvm", "-o", compiled_file, source] + compiler_options
         subprocess.check_call(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        
+
         compiled_library =  os.path.join(instance_dir, "library.bc")
         include_path=os.path.join(self.include_path, "klee-test-comp.c")
         cmd = [path_to_clang, "-c", "-g", "-emit-llvm", "-o", compiled_library, include_path] + compiler_options
@@ -124,8 +124,7 @@ class TestRunner(object):
         for filename in os.listdir(self.tmp):
             if filename.startswith("GCOV_coverage"):
                 shutil.copy(os.path.join(self.tmp,filename), cur_dir)
-    
-        
+
 
 def main():
     parser = argparse.ArgumentParser()
@@ -135,6 +134,7 @@ def main():
     parser.add_argument("--output", required=True, help="where to keep coverage output, required")
     parser.add_argument("--builddir", default=None, help="build directory, default: build")
     parser.add_argument("--llvm", default=None)
+    parser.add_argument("--just-compile", choices=('True','False'), help="Just compile files")
     args = parser.parse_args()
 
     wrapper = TestRunner(args.source, args.t, args.klee_output, args.output, args.builddir, args.llvm)
@@ -142,14 +142,17 @@ def main():
     num_sources = len([name for name in os.listdir(args.source)])
     print(f"{num_sources} total.")
     for filename in os.listdir(args.source):
-        print(f"Source {count}/{num_sources}")
-        wrapper.compile(filename)
-        wrapper.run(filename)
-        wrapper.compile_coverage(filename)
-        wrapper.run_coverage(filename)
-        count += 1
-    wrapper.save_gcov()
-    shutil.rmtree(wrapper.tmp)
+        if args.just_compile == 'True':
+            wrapper.compile(filename)
+        else:
+            print(f"Source {count}/{num_sources}")
+            wrapper.compile(filename)
+            wrapper.run(filename)
+            wrapper.compile_coverage(filename)
+            wrapper.run_coverage(filename)
+            count += 1
+            wrapper.save_gcov()
+            shutil.rmtree(wrapper.tmp)
 
 if __name__ == '__main__':
     main()
